+++ Utility/Diagnostics/DiagFramePlane.cpp
@@ -220,7 +220,7 @@ DiagFramePlane::processDiag(
   amrex::Vector<amrex::MultiFab> planeData(a_state.size());
   for (int lev = 0; lev < a_state.size(); ++lev) {
     planeData[lev].define(
-      m_sliceBA[lev], m_sliceDM[lev], m_fieldNames.size(), 0);
+      m_sliceBA[lev], m_sliceDM[lev], static_cast<int>(m_fieldNames.size()), 0);
     int p0 = m_k0[lev];
 #ifdef AMREX_USE_OMP
 #pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
